requirements.txt
streamlit>=1.42.0
requests>=2.31.0
python-dateutil>=2.8.2
